changed data type from int to uint8_t

+++ ATMEGA4809_AutomatedProject.X/newmain.c
@@ -8,18 +8,18 @@
 
 #include <avr/io.h>
 
-int m =10;
+uint8_t m =10;
 int add(uint8_t n, uint8_t o, uint8_t p ){
    return n+o+p;
 }
 int main() {
-   volatile int a = 0x05;
+   volatile uint8_t a = 0x05;
    PORTA.DIR = a;
-   volatile int c = 0x1F;
+   volatile uint8_t c = 0x1F;
    PORTB.DIR = c;
-   volatile int b =0x08;
+   volatile uint8_t b =0x08;
    PORTC.DIR = b;
-   volatile int t = add(a,b,c);
+   volatile uint8_t t = add(a,b,c);
    PORTD.DIR =t;
    return 0;
 }
